requirements.txt
pillow==5.4.1
pyinstaller==3.4
pyqt5==5.12.1
python-docx==0.8.10
